requirements.txt
flask==2.3.2
flask-sqlalchemy==3.0.3
Pillow==9.5.0
numpy==1.24.3
matplotlib==3.7.1
opencv-python==4.7.0.72
pandas==2.0.1
